Refactor CSV import
Move CSV import related functions into import-csv.c.
